swag_Automation_QA/requirements.txt
anyio==4.4.0
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
comtypes==1.4.8
cryptography==43.0.0
decorator==5.1.1
Faker==27.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpath-ng==1.7.0
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyboard==0.13.5
MouseInfo==0.1.3
multi_key_dict==2.0.3
objectpath==0.6.1
outcome==1.3.0.post0
packaging==24.1
pbr==6.0.0
pillow==10.4.0
ply==3.11
prompt_toolkit==3.0.47
psutil==5.9.8
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pyjsonselect==0.2.2
PyMsgBox==1.0.9
pyobjc==10.3.2
pyobjc-core==10.3.2
pyobjc-framework-Accessibility==10.3.2
pyobjc-framework-Accounts==10.3.2
pyobjc-framework-AddressBook==10.3.2
pyobjc-framework-AdServices==10.3.2
pyobjc-framework-AdSupport==10.3.2
pyobjc-framework-AppleScriptKit==10.3.2
pyobjc-framework-AppleScriptObjC==10.3.2
pyobjc-framework-ApplicationServices==10.3.2
pyobjc-framework-AppTrackingTransparency==10.3.2
pyobjc-framework-AudioVideoBridging==10.3.2
pyobjc-framework-AuthenticationServices==10.3.2
pyobjc-framework-AutomaticAssessmentConfiguration==10.3.2
pyobjc-framework-Automator==10.3.2
pyobjc-framework-AVFoundation==10.3.2
pyobjc-framework-AVKit==10.3.2
pyobjc-framework-AVRouting==10.3.2
pyobjc-framework-BackgroundAssets==10.3.2
pyobjc-framework-BrowserEngineKit==10.3.2
pyobjc-framework-BusinessChat==10.3.2
pyobjc-framework-CalendarStore==10.3.2
pyobjc-framework-CallKit==10.3.2
pyobjc-framework-CFNetwork==10.3.2
pyobjc-framework-Cinematic==10.3.2
pyobjc-framework-ClassKit==10.3.2
pyobjc-framework-CloudKit==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Collaboration==10.3.2
pyobjc-framework-ColorSync==10.3.2
pyobjc-framework-Contacts==10.3.2
pyobjc-framework-ContactsUI==10.3.2
pyobjc-framework-CoreAudio==10.3.2
pyobjc-framework-CoreAudioKit==10.3.2
pyobjc-framework-CoreBluetooth==10.3.2
pyobjc-framework-CoreData==10.3.2
pyobjc-framework-CoreHaptics==10.3.2
pyobjc-framework-CoreLocation==10.3.2
pyobjc-framework-CoreMedia==10.3.2
pyobjc-framework-CoreMediaIO==10.3.2
pyobjc-framework-CoreMIDI==10.3.2
pyobjc-framework-CoreML==10.3.2
pyobjc-framework-CoreMotion==10.3.2
pyobjc-framework-CoreServices==10.3.2
pyobjc-framework-CoreSpotlight==10.3.2
pyobjc-framework-CoreText==10.3.2
pyobjc-framework-CoreWLAN==10.3.2
pyobjc-framework-CryptoTokenKit==10.3.2
pyobjc-framework-DataDetection==10.3.2
pyobjc-framework-DeviceCheck==10.3.2
pyobjc-framework-DictionaryServices==10.3.2
pyobjc-framework-DiscRecording==10.3.2
pyobjc-framework-DiscRecordingUI==10.3.2
pyobjc-framework-DiskArbitration==10.3.2
pyobjc-framework-DVDPlayback==10.3.2
pyobjc-framework-EventKit==10.3.2
pyobjc-framework-ExceptionHandling==10.3.2
pyobjc-framework-ExecutionPolicy==10.3.2
pyobjc-framework-ExtensionKit==10.3.2
pyobjc-framework-ExternalAccessory==10.3.2
pyobjc-framework-FileProvider==10.3.2
pyobjc-framework-FileProviderUI==10.3.2
pyobjc-framework-FinderSync==10.3.2
pyobjc-framework-FSEvents==10.3.2
pyobjc-framework-GameCenter==10.3.2
pyobjc-framework-GameController==10.3.2
pyobjc-framework-GameKit==10.3.2
pyobjc-framework-GameplayKit==10.3.2
pyobjc-framework-HealthKit==10.3.2
pyobjc-framework-ImageCaptureCore==10.3.2
pyobjc-framework-InputMethodKit==10.3.2
pyobjc-framework-InstallerPlugins==10.3.2
pyobjc-framework-InstantMessage==10.3.2
pyobjc-framework-Intents==10.3.2
pyobjc-framework-IntentsUI==10.3.2
pyobjc-framework-IOBluetooth==10.3.2
pyobjc-framework-IOBluetoothUI==10.3.2
pyobjc-framework-IOSurface==10.3.2
pyobjc-framework-iTunesLibrary==10.3.2
pyobjc-framework-KernelManagement==10.3.2
pyobjc-framework-LatentSemanticMapping==10.3.2
pyobjc-framework-LaunchServices==10.3.2
pyobjc-framework-libdispatch==10.3.2
pyobjc-framework-libxpc==10.3.2
pyobjc-framework-LinkPresentation==10.3.2
pyobjc-framework-LocalAuthentication==10.3.2
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.3.2
pyobjc-framework-MailKit==10.3.2
pyobjc-framework-MapKit==10.3.2
pyobjc-framework-MediaAccessibility==10.3.2
pyobjc-framework-MediaLibrary==10.3.2
pyobjc-framework-MediaPlayer==10.3.2
pyobjc-framework-MediaToolbox==10.3.2
pyobjc-framework-Metal==10.3.2
pyobjc-framework-MetalFX==10.3.2
pyobjc-framework-MetalKit==10.3.2
pyobjc-framework-MetalPerformanceShaders==10.3.2
pyobjc-framework-MetalPerformanceShadersGraph==10.3.2
pyobjc-framework-MetricKit==10.3.2
pyobjc-framework-MLCompute==10.3.2
pyobjc-framework-ModelIO==10.3.2
pyobjc-framework-MultipeerConnectivity==10.3.2
pyobjc-framework-NaturalLanguage==10.3.2
pyobjc-framework-NetFS==10.3.2
pyobjc-framework-Network==10.3.2
pyobjc-framework-NetworkExtension==10.3.2
pyobjc-framework-NotificationCenter==10.3.2
pyobjc-framework-OpenDirectory==10.3.2
pyobjc-framework-OSAKit==10.3.2
pyobjc-framework-OSLog==10.3.2
pyobjc-framework-PassKit==10.3.2
pyobjc-framework-PencilKit==10.3.2
pyobjc-framework-PHASE==10.3.2
pyobjc-framework-Photos==10.3.2
pyobjc-framework-PhotosUI==10.3.2
pyobjc-framework-PreferencePanes==10.3.2
pyobjc-framework-PushKit==10.3.2
pyobjc-framework-Quartz==10.3.2
pyobjc-framework-QuickLookThumbnailing==10.3.2
pyobjc-framework-ReplayKit==10.3.2
pyobjc-framework-SafariServices==10.3.2
pyobjc-framework-SafetyKit==10.3.2
pyobjc-framework-SceneKit==10.3.2
pyobjc-framework-ScreenCaptureKit==10.3.2
pyobjc-framework-ScreenSaver==10.3.2
pyobjc-framework-ScreenTime==10.3.2
pyobjc-framework-ScriptingBridge==10.3.2
pyobjc-framework-SearchKit==10.3.2
pyobjc-framework-Security==10.3.2
pyobjc-framework-SecurityFoundation==10.3.2
pyobjc-framework-SecurityInterface==10.3.2
pyobjc-framework-SensitiveContentAnalysis==10.3.2
pyobjc-framework-ServiceManagement==10.3.2
pyobjc-framework-SharedWithYou==10.3.2
pyobjc-framework-SharedWithYouCore==10.3.2
pyobjc-framework-ShazamKit==10.3.2
pyobjc-framework-Social==10.3.2
pyobjc-framework-SoundAnalysis==10.3.2
pyobjc-framework-Speech==10.3.2
pyobjc-framework-SpriteKit==10.3.2
pyobjc-framework-StoreKit==10.3.2
pyobjc-framework-Symbols==10.3.2
pyobjc-framework-SyncServices==10.3.2
pyobjc-framework-SystemConfiguration==10.3.2
pyobjc-framework-SystemExtensions==10.3.2
pyobjc-framework-ThreadNetwork==10.3.2
pyobjc-framework-UniformTypeIdentifiers==10.3.2
pyobjc-framework-UserNotifications==10.3.2
pyobjc-framework-UserNotificationsUI==10.3.2
pyobjc-framework-VideoSubscriberAccount==10.3.2
pyobjc-framework-VideoToolbox==10.3.2
pyobjc-framework-Virtualization==10.3.2
pyobjc-framework-Vision==10.3.2
pyobjc-framework-WebKit==10.3.2
pyOpenSSL==24.2.1
pyotp==2.9.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jenkins==1.8.2
python-telegram-bot==21.4
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
robotframework==7.0.1
robotframework-datetime-tz==1.0.6
robotframework-debuglibrary==2.5.0
robotframework-faker==5.0.0
robotframework-jsonlibrary==0.5
robotframework-jsonvalidator==2.0.0
robotframework-pythonlibcore==4.4.1
robotframework-randomlibrary==0.0.2
robotframework-requests==0.9.7
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==6.3.0
robotframework-yamllibrary==0.2.8
rpaframework-core==11.3.3
rpds-py==0.20.0
rubicon-objc==0.4.9
selenium==4.23.1
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
uiautomation==2.0.20
urllib3==2.2.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
